moved some utility functions out
